Throw the error IDs as UINT to be sure they are correctly caught. This commit fixes ticket #2355.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@5010 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/mpc-hc/MainFrm.cpp
@@ -10379,7 +10379,7 @@ void CMainFrame::OpenCreateGraphObject(OpenMediaData* pOMD)
 
 			pUnk = (IUnknown*)(INonDelegatingUnknown*)DNew CRealMediaGraph(m_pVideoWnd->m_hWnd, hr);
 			if (!pUnk) {
-				throw IDS_AG_OUT_OF_MEMORY;
+				throw (UINT)IDS_AG_OUT_OF_MEMORY;
 			}
 
 			if (SUCCEEDED(hr)) {
@@ -10391,14 +10391,14 @@ void CMainFrame::OpenCreateGraphObject(OpenMediaData* pOMD)
 		} else if (engine == ShockWave) {
 			pUnk = (IUnknown*)(INonDelegatingUnknown*)DNew CShockwaveGraph(m_pVideoWnd->m_hWnd, hr);
 			if (!pUnk) {
-				throw IDS_AG_OUT_OF_MEMORY;
+				throw (UINT)IDS_AG_OUT_OF_MEMORY;
 			}
 
 			if (SUCCEEDED(hr)) {
 				pGB = CComQIPtr<IGraphBuilder>(pUnk);
 			}
 			if (FAILED(hr) || !pGB) {
-				throw IDS_MAINFRM_77;
+				throw (UINT)IDS_MAINFRM_77;
 			}
 			m_fShockwaveGraph = true;
 		} else if (engine == QuickTime) {
@@ -10407,7 +10407,7 @@ void CMainFrame::OpenCreateGraphObject(OpenMediaData* pOMD)
 #else
 			pUnk = (IUnknown*)(INonDelegatingUnknown*)DNew CQuicktimeGraph(m_pVideoWnd->m_hWnd, hr);
 			if (!pUnk) {
-				throw IDS_AG_OUT_OF_MEMORY;
+				throw (UINT)IDS_AG_OUT_OF_MEMORY;
 			}
 
 			if (SUCCEEDED(hr)) {
@@ -10435,7 +10435,7 @@ void CMainFrame::OpenCreateGraphObject(OpenMediaData* pOMD)
 	}
 
 	if (!pGB) {
-		throw IDS_MAINFRM_80;
+		throw (UINT)IDS_MAINFRM_80;
 	}
 
 	pGB->AddToROT();
@@ -10451,11 +10451,11 @@ void CMainFrame::OpenCreateGraphObject(OpenMediaData* pOMD)
 	if (!(pMC && pME && pMS)
 			|| !(pVW && pBV)
 			|| !(pBA)) {
-		throw IDS_GRAPH_INTERFACES_ERROR;
+		throw (UINT)IDS_GRAPH_INTERFACES_ERROR;
 	}
 
 	if (FAILED(pME->SetNotifyWindow((OAHWND)m_hWnd, WM_GRAPHNOTIFY, 0))) {
-		throw IDS_GRAPH_TARGET_WND_ERROR;
+		throw (UINT)IDS_GRAPH_TARGET_WND_ERROR;
 	}
 
 	m_pProv = (IUnknown*)DNew CKeyProvider();
@@ -10480,7 +10480,7 @@ CWnd *CMainFrame::GetModalParent()
 void CMainFrame::OpenFile(OpenFileData* pOFD)
 {
 	if (pOFD->fns.IsEmpty()) {
-		throw IDS_MAINFRM_81;
+		throw (UINT)IDS_MAINFRM_81;
 	}
 
 	AppSettings& s = AfxGetAppSettings();
@@ -10775,17 +10775,17 @@ void CMainFrame::OpenDVD(OpenDVDData* pODD)
 	EndEnumFilters;
 
 	if (hr == E_INVALIDARG) {
-		throw IDS_MAINFRM_93;
+		throw (UINT)IDS_MAINFRM_93;
 	} else if (hr == VFW_E_CANNOT_RENDER) {
-		throw IDS_DVD_NAV_ALL_PINS_ERROR;
+		throw (UINT)IDS_DVD_NAV_ALL_PINS_ERROR;
 	} else if (hr == VFW_S_PARTIAL_RENDER) {
-		throw IDS_DVD_NAV_SOME_PINS_ERROR;
+		throw (UINT)IDS_DVD_NAV_SOME_PINS_ERROR;
 	} else if (hr == E_NOINTERFACE || !pDVDC || !pDVDI) {
-		throw IDS_DVD_INTERFACES_ERROR;
+		throw (UINT)IDS_DVD_INTERFACES_ERROR;
 	} else if (hr == VFW_E_CANNOT_LOAD_SOURCE_FILTER) {
-		throw IDS_MAINFRM_94;
+		throw (UINT)IDS_MAINFRM_94;
 	} else if (FAILED(hr)) {
-		throw IDS_AG_FAILED;
+		throw (UINT)IDS_AG_FAILED;
 	}
 
 	WCHAR buff[_MAX_PATH];
@@ -10832,7 +10832,7 @@ void CMainFrame::OpenCapture(OpenDeviceData* pODD)
 
 	if (!m_VidDispName.IsEmpty()) {
 		if (!CreateFilter(m_VidDispName, &pVidCapTmp, vidfrname)) {
-			throw IDS_MAINFRM_96;
+			throw (UINT)IDS_MAINFRM_96;
 		}
 	}
 
@@ -10840,12 +10840,12 @@ void CMainFrame::OpenCapture(OpenDeviceData* pODD)
 
 	if (!m_AudDispName.IsEmpty()) {
 		if (!CreateFilter(m_AudDispName, &pAudCapTmp, audfrname)) {
-			throw IDS_MAINFRM_96;
+			throw (UINT)IDS_MAINFRM_96;
 		}
 	}
 
 	if (!pVidCapTmp && !pAudCapTmp) {
-		throw IDS_MAINFRM_98;
+		throw (UINT)IDS_MAINFRM_98;
 	}
 
 	pCGB = NULL;
@@ -10853,7 +10853,7 @@ void CMainFrame::OpenCapture(OpenDeviceData* pODD)
 	pAudCap = NULL;
 
 	if (FAILED(pCGB.CoCreateInstance(CLSID_CaptureGraphBuilder2))) {
-		throw IDS_MAINFRM_99;
+		throw (UINT)IDS_MAINFRM_99;
 	}
 
 	HRESULT hr;
@@ -10862,7 +10862,7 @@ void CMainFrame::OpenCapture(OpenDeviceData* pODD)
 
 	if (pVidCapTmp) {
 		if (FAILED(hr = pGB->AddFilter(pVidCapTmp, vidfrname))) {
-			throw IDS_CAPTURE_ERROR_VID_FILTER;
+			throw (UINT)IDS_CAPTURE_ERROR_VID_FILTER;
 		}
 
 		pVidCap = pVidCapTmp;
@@ -10941,7 +10941,7 @@ void CMainFrame::OpenCapture(OpenDeviceData* pODD)
 
 	if (pAudCapTmp) {
 		if (FAILED(hr = pGB->AddFilter(pAudCapTmp, CStringW(audfrname)))) {
-			throw IDS_CAPTURE_ERROR_AUD_FILTER;
+			throw (UINT)IDS_CAPTURE_ERROR_AUD_FILTER;
 		}
 
 		pAudCap = pAudCapTmp;
@@ -10972,7 +10972,7 @@ void CMainFrame::OpenCapture(OpenDeviceData* pODD)
 	}
 
 	if (!(pVidCap || pAudCap)) {
-		throw IDS_MAINFRM_108;
+		throw (UINT)IDS_MAINFRM_108;
 	}
 
 	pODD->title.LoadString(IDS_CAPTURE_LIVE);
@@ -11641,19 +11641,19 @@ bool CMainFrame::OpenMediaPrivate(CAutoPtr<OpenMediaData> pOMD)
 		CComPtr<IMFVideoMixerBitmap>	pMFVMB;
 		CComPtr<IMadVRTextOsd>			pMVTO;
 		if (m_fOpeningAborted) {
-			throw IDS_AG_ABORTED;
+			throw (UINT)IDS_AG_ABORTED;
 		}
 
 		OpenCreateGraphObject(pOMD);
 
 		if (m_fOpeningAborted) {
-			throw IDS_AG_ABORTED;
+			throw (UINT)IDS_AG_ABORTED;
 		}
 
 		SetupIViAudReg();
 
 		if (m_fOpeningAborted) {
-			throw IDS_AG_ABORTED;
+			throw (UINT)IDS_AG_ABORTED;
 		}
 
 		if (pFileData) {
@@ -11664,13 +11664,13 @@ bool CMainFrame::OpenMediaPrivate(CAutoPtr<OpenMediaData> pOMD)
 			if (s.iDefaultCaptureDevice == 1) {
 				HRESULT hr = OpenBDAGraph();
 				if (FAILED(hr)) {
-					throw IDS_CAPTURE_ERROR_DEVICE;
+					throw (UINT)IDS_CAPTURE_ERROR_DEVICE;
 				}
 			} else {
 				OpenCapture(pDeviceData);
 			}
 		} else {
-			throw IDS_INVALID_PARAMS_ERROR;
+			throw (UINT)IDS_INVALID_PARAMS_ERROR;
 		}
 
 		m_pCAP2 = NULL;
@@ -11728,25 +11728,25 @@ bool CMainFrame::OpenMediaPrivate(CAutoPtr<OpenMediaData> pOMD)
 		EndEnumFilters;
 
 		if (m_fOpeningAborted) {
-			throw IDS_AG_ABORTED;
+			throw (UINT)IDS_AG_ABORTED;
 		}
 
 		OpenCustomizeGraph();
 
 		if (m_fOpeningAborted) {
-			throw IDS_AG_ABORTED;
+			throw (UINT)IDS_AG_ABORTED;
 		}
 
 		OpenSetupVideo();
 
 		if (m_fOpeningAborted) {
-			throw IDS_AG_ABORTED;
+			throw (UINT)IDS_AG_ABORTED;
 		}
 
 		OpenSetupAudio();
 
 		if (m_fOpeningAborted) {
-			throw IDS_AG_ABORTED;
+			throw (UINT)IDS_AG_ABORTED;
 		}
 
 		if (m_pCAP && (!m_fAudioOnly || m_fRealMediaGraph)) {
@@ -11791,7 +11791,7 @@ bool CMainFrame::OpenMediaPrivate(CAutoPtr<OpenMediaData> pOMD)
 		}
 
 		if (m_fOpeningAborted) {
-			throw IDS_AG_ABORTED;
+			throw (UINT)IDS_AG_ABORTED;
 		}
 
 		OpenSetupWindowTitle(pOMD->title);
